move component in foundations and component folder
