+++ src/components/CodeInput/CodeInput.tsx
@@ -1,6 +1,6 @@
 import { Box, Stack, Typography, useTheme, styled, keyframes } from '@mui/material';
 import { useEffect, useId, useLayoutEffect, useRef, useState } from 'react';
-import { blue, error as errorColor, neutral as neutralColor } from './../../theme/colors';
+import { palette } from '../../theme/foundations/colors';
 
 /**
  * Layout constants used to size the code input component.
@@ -72,7 +72,7 @@ const Caret = styled('div')<{ position: CaretPosition['position'] }>(({ position
   width: '2px',
   borderRadius: '2px',
   height: '1em',
-  backgroundColor: blue[500],
+  backgroundColor: palette.blue[500],
   animation: `${blink} 1s step-start infinite`,
   transformOrigin: 'center',
   left: position === 'center' ? '50%' : position === 'end' ? 'calc(100% + 1px)' : '-1px',
@@ -87,7 +87,7 @@ const CodeBox = styled(Box, {
     codeBoxPaddingBottom
   )}`,
   border: `${theme.spacing(error ? codeBoxErrorBorder : 0.125)} solid ${
-    error ? errorColor[600] : neutralColor[100]
+    error ? palette.error[600] : palette.neutral.grey[100]
   }`,
   borderRadius: theme.spacing(codeBoxBorderRadius),
 }));
@@ -98,7 +98,7 @@ const CharBox = styled(Box)(({ theme }) => ({
   lineHeight: '1.5em',
   paddingBottom: theme.spacing(0.25),
   marginBottom: theme.spacing(0.5),
-  borderBottom: `1px solid ${neutralColor[700]}`,
+  borderBottom: `1px solid ${palette.neutral.grey[700]}`,
   display: 'flex',
   alignItems: 'center',
   justifyContent: 'center',
@@ -119,7 +119,7 @@ const HelperText = styled(Typography, {
   boxSizing: 'content-box',
   paddingLeft: theme.spacing(3),
   paddingRight: theme.spacing(3),
-  color: error ? errorColor[600] : theme.palette.text.primary,
+  color: error ? palette.error[600] : theme.palette.text.primary,
 }));
 
 /**
