+++ src/theme/components/icon.ts
@@ -1,0 +1,13 @@
+import { palette } from '../foundations/colors';
+import { opacity } from '../foundations/opacity';
+
+export const icon = {
+  active: palette.neutral.black,
+  decorative: palette.neutral.grey[300],
+  action: palette.blue[500],
+  success: palette.success[700],
+  info: palette.info[700],
+  warning: palette.warning[700],
+  error: palette.error[600],
+  button: opacity.black[60],
+};
