+++ src/theme/components/progress.ts
@@ -1,0 +1,6 @@
+import { palette } from 'theme/foundations/colors';
+
+export const progress = {
+  track: palette.blue[500],
+  total: palette.neutral.white,
+};
