+++ src/theme/foundations/divider.ts
@@ -1,0 +1,5 @@
+import { palette } from './colors';
+
+export const divider = {
+  default: palette.neutral.grey[100],
+};
