+++ src/theme/foundations/surface.ts
@@ -1,0 +1,12 @@
+import { palette } from './colors';
+import { opacity } from './opacity';
+
+export const surface = {
+  primary: palette.neutral.white,
+  overlay: opacity.black[20],
+  blue: palette.blue[500],
+  skeleton: palette.neutral.grey[100],
+  skeletonOpacity: palette.neutral.grey[450],
+  placeholder: palette.neutral.grey[50],
+  blueItalia: palette.blueitalia[50],
+};
